refactor(paste): move output.paste → os_/paste/
