+++ src/dicton/os_/paste/_darwin.py
@@ -1,0 +1,19 @@
+"""macOS paste backend: pbcopy + AppleScript Cmd+V."""
+
+from __future__ import annotations
+
+import subprocess
+import time
+
+
+def paste_darwin(text: str) -> None:
+    subprocess.run(["pbcopy"], input=text.encode("utf-8"), check=True)
+    time.sleep(0.02)
+    subprocess.run(
+        [
+            "osascript",
+            "-e",
+            'tell application "System Events" to keystroke "v" using command down',
+        ],
+        check=False,
+    )
